+++ packages/fastboard-react/test/main.tsx
@@ -39,10 +39,10 @@ function App() {
     },
     joinRoom: {
       uid: genUID(),
-      uuid: import.meta.env.VITE_ROOM_UUID || "1bfe7360a23d11ef844c9de7e0348444",
+      uuid: import.meta.env.VITE_ROOM_UUID || "f3e75f20c0f211ef940ec9f75538d484",
       roomToken:
         import.meta.env.VITE_ROOM_TOKEN ||
-        "NETLESSROOM_YWs9VWtNUk92M1JIN2I2Z284dCZleHBpcmVBdD0xNzMxNjQzMjgyMDQyJm5vbmNlPTFjMjEzZGEwLWEyM2QtMTFlZi05NmE5LWFiMzg4NjE4OThhZiZyb2xlPTEmc2lnPWM2MzYwMDU5MGRkOTY4NmJkYzEwMzYxNWEwYTJhODY1OWY1ZjFjNjZmNTRkY2FiMzVkMzE5YzQ4ZTk5YmIzOTYmdXVpZD0xYmZlNzM2MGEyM2QxMWVmODQ0YzlkZTdlMDM0ODQ0NA",
+        "NETLESSROOM_YWs9VWtNUk92M1JIN2I2Z284dCZleHBpcmVBdD0xNzM1MDE5OTE4MDM3Jm5vbmNlPWY0MDUyMDUwLWMwZjItMTFlZi05NmE5LWFiMzg4NjE4OThhZiZyb2xlPTEmc2lnPWZmNmQwNzU5NDJlNTc2ZGVhZWMwMzM5ZjUwYTkzMmNlM2M3MTBlZmRlNzczYjExMGE5YjMwOGRkMzUyNzM4MDkmdXVpZD1mM2U3NWYyMGMwZjIxMWVmOTQwZWM5Zjc1NTM4ZDQ4NA",
       isWritable: search.get("isWritable") !== "0",
       floatBar: true,
     },