+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
+## 1.0.2 (2024-12-23)
+- When turning on appliance plugin, Fastboard UI added `markPen`.
 
 ## 1.0.1 (2024-12-17)
 - Fixed when import `@netless/fastboard-ui/full` the declaration file `full.d.ts` lost .
 
 ## 1.0.0 (2024-12-02)
 
-- When turning on appliance plugin, Fastboard UI add `laserPen`.
+- When turning on appliance plugin, Fastboard UI added `laserPen`.
 - update `@netless/window-manager` to `^1.0.0`.
 
 ## 0.3.22 (2024-11-05)